ISSUE #2169 - Fixed showing all meshes when the frame was loading
